Move IBLayoutConstraintBuilder to dedicated file
