+++ Sources/UIViewKit/UIViewDSL/UIViewDSL+ResultBuilders.swift
@@ -40,39 +40,3 @@ public enum IBSubviewsBuilder {
         components.flatMap { $0 }
     }
 }
-
-@resultBuilder
-public enum IBLayoutConstraintBuilder {
-
-    public static func buildBlock(_ components: [NSLayoutConstraint]...) -> [NSLayoutConstraint] {
-        components.flatMap { $0 }
-    }
-
-    public static func buildBlock(_ components: NSLayoutConstraint...) -> [NSLayoutConstraint] {
-        components
-    }
-
-    public static func buildOptional(_ component: [NSLayoutConstraint]?) -> [NSLayoutConstraint] {
-        component ?? []
-    }
-
-    public static func buildEither(first component: [NSLayoutConstraint]) -> [NSLayoutConstraint] {
-        component
-    }
-
-    public static func buildEither(second component: [NSLayoutConstraint]) -> [NSLayoutConstraint] {
-        component
-    }
-
-    public static func buildExpression(_ expression: NSLayoutConstraint) -> [NSLayoutConstraint] {
-        [expression]
-    }
-
-    public static func buildExpression(_ expression: [NSLayoutConstraint]) -> [NSLayoutConstraint] {
-        expression
-    }
-
-    public static func buildExpression(_ expression: Any?) -> [NSLayoutConstraint] {
-        []
-    }
-}
